move more things to helpers
